refactor(analyze): extract formatters.ts and github.ts from index.ts (#8)
- Create tools/analyze/formatters.ts: ANSI color palette, severity helpers,
  formatBar progress bar, renderTerminalReport, renderJsonReport
- Create tools/analyze/github.ts: LABEL_COLORS, ensureLabels, buildIssueData,
  createGitHubIssue
- Reduce tools/analyze/index.ts from 641 to 375 lines by importing from the
  two new modules; thin orchestrator retains analyze(), parseAnalyzeFlags(),
  and REFACTOR_HELP
- All 8 analyze modules now created and integrated (types, language-profiles,
  cache, discovery, tier1, tier2, formatters, github)
