+++ tools/analyze/index.ts
@@ -44,280 +44,15 @@ import type {
 	Tier2Result,
 	AnalysisResult,
 	RefactorReport,
-	IssueData,
 } from './types.ts';
 import { computeFileHash, loadCache, saveCache } from './cache.ts';
 import { discoverFiles } from './discovery.ts';
 import { analyzeTier1 } from './tier1.ts';
 import { analyzeTier2 } from './tier2.ts';
+import { COLORS, renderTerminalReport, renderJsonReport } from './formatters.ts';
+import { buildIssueData, createGitHubIssue, ensureLabels } from './github.ts';
 
 export type { RefactorFlags } from './types.ts';
-
-// ─── GitHub Label Management ────────────────────────────────────────────────
-
-const LABEL_COLORS: Record<string, string> = {
-	'refactor': '1d76db',
-	'ai-suggested': 'c5def5',
-	'priority:high': 'e11d48',
-};
-
-async function ensureLabels(labels: string[], repoRoot: string): Promise<void> {
-	for (const label of labels) {
-		const check = Bun.spawnSync(['gh', 'label', 'list', '--search', label, '--json', 'name'], {
-			cwd: repoRoot,
-			stdout: 'pipe',
-			stderr: 'pipe',
-		});
-
-		const output = new TextDecoder().decode(check.stdout as Buffer).trim();
-		let exists = false;
-		try {
-			const parsed = JSON.parse(output) as { name: string }[];
-			exists = parsed.some(l => l.name === label);
-		} catch {}
-
-		if (!exists) {
-			const color = LABEL_COLORS[label] ?? 'ededed';
-			const create = Bun.spawnSync(
-				['gh', 'label', 'create', label, '--color', color, '--force'],
-				{ cwd: repoRoot, stdout: 'pipe', stderr: 'pipe' }
-			);
-			if (create.exitCode === 0) {
-				log.info(`Created missing label: ${label}`);
-			} else {
-				log.warn(`Could not create label '${label}' — issue creation may fail`);
-			}
-		}
-	}
-}
-
-// ─── GitHub Issue Creation ──────────────────────────────────────────────────
-
-function buildIssueData(result: AnalysisResult): IssueData | null {
-	if (result.tier1.severity === 'ok' && !result.tier2) return null;
-
-	const { tier1, tier2 } = result;
-	const splitCount = tier2?.suggestions?.length ?? 0;
-
-	const title = splitCount > 0
-		? `refactor(${tier1.relativePath}): split into ${splitCount} focused modules`
-		: `refactor(${tier1.relativePath}): decompose file (${tier1.lineCount} lines)`;
-
-	let body = `## File Structure Analysis\n\n`;
-	body += `**File:** \`${tier1.relativePath}\`\n`;
-	body += `**Language:** ${tier1.language}\n`;
-	body += `**Lines:** ${tier1.lineCount} (soft: ${tier1.softThreshold}, hard: ${tier1.hardThreshold})\n`;
-	body += `**Severity:** ${tier1.severity}\n\n`;
-
-	body += `### Heuristic Signals\n\n`;
-	for (const signal of tier1.signals) {
-		body += `- ${signal}\n`;
-	}
-
-	if (tier2) {
-		body += `\n### Detected Responsibilities\n\n`;
-		for (const r of tier2.responsibilities) {
-			body += `- **${r.name}**: ${r.description} (lines ${r.lineRanges})\n`;
-		}
-
-		if (tier2.suggestions.length > 0) {
-			body += `\n### Suggested Split\n\n`;
-			for (const s of tier2.suggestions) {
-				body += `- \`${s.filename}\` — ${s.responsibilities.join(', ')}\n`;
-				body += `  - _${s.rationale}_\n`;
-			}
-		}
-
-		if (tier2.principles.length > 0) {
-			body += `\n### Principle Violations\n\n`;
-			for (const p of tier2.principles) {
-				body += `- ${p}\n`;
-			}
-		}
-
-		body += `\n### Effort: ${tier2.effort}\n\n`;
-		body += `${tier2.summary}\n`;
-
-		// Acceptance criteria from Tier 2 analysis
-		body += `\n### Acceptance Criteria\n\n`;
-		if (tier2.suggestions.length > 0) {
-			for (const s of tier2.suggestions) {
-				body += `- [ ] \`${s.filename}\` created with ${s.responsibilities.join(', ')} responsibility\n`;
-			}
-		}
-		body += `- [ ] All existing exports re-exported or migrated (no broken imports)\n`;
-		body += `- [ ] Original file removed or reduced to re-exports only\n`;
-		body += `- [ ] No resulting file exceeds ${tier1.softThreshold} lines (soft threshold)\n`;
-		body += `- [ ] Tests pass\n`;
-	} else {
-		// No Tier 2 data — flag for manual scoping
-		body += `\n> **Note:** This issue was generated from heuristic signals only (no AI analysis). `;
-		body += `It may need manual scoping or a \`pait analyze --budget\` re-run before automated implementation.\n`;
-
-		body += `\n### Acceptance Criteria\n\n`;
-		body += `- [ ] File decomposed into focused modules with single responsibilities\n`;
-		body += `- [ ] All existing exports re-exported or migrated (no broken imports)\n`;
-		body += `- [ ] No resulting file exceeds ${tier1.softThreshold} lines (soft threshold)\n`;
-		body += `- [ ] Tests pass\n`;
-	}
-
-	body += `\n---\n_Generated by \`pait analyze\`_\n`;
-
-	const labels = ['refactor', 'ai-suggested'];
-	if (tier1.severity === 'critical') labels.push('priority:high');
-
-	return { title, body, labels };
-}
-
-async function createGitHubIssue(issue: IssueData, repoRoot: string, dryRun: boolean): Promise<string | null> {
-	if (dryRun) {
-		log.info(`[DRY RUN] Would create: ${issue.title}`);
-		return null;
-	}
-
-	const labelArgs = issue.labels.flatMap(l => ['--label', l]);
-
-	const proc = Bun.spawn(
-		['gh', 'issue', 'create', '--title', issue.title, '--body', issue.body, ...labelArgs],
-		{
-			cwd: repoRoot,
-			stdout: 'pipe',
-			stderr: 'pipe',
-		}
-	);
-
-	const output = await new Response(proc.stdout).text();
-	const exitCode = await proc.exited;
-
-	if (exitCode !== 0) {
-		const stderr = await new Response(proc.stderr).text();
-		log.warn(`Failed to create issue: ${stderr.trim()}`);
-		return null;
-	}
-
-	return output.trim();
-}
-
-// ─── Output Formatters ──────────────────────────────────────────────────────
-
-const COLORS = {
-	reset: '\x1b[0m',
-	bold: '\x1b[1m',
-	dim: '\x1b[2m',
-	red: '\x1b[31m',
-	green: '\x1b[32m',
-	yellow: '\x1b[33m',
-	blue: '\x1b[34m',
-	magenta: '\x1b[35m',
-	cyan: '\x1b[36m',
-	white: '\x1b[37m',
-	bgRed: '\x1b[41m',
-	bgYellow: '\x1b[43m',
-	bgGreen: '\x1b[42m',
-};
-
-function severityColor(severity: 'ok' | 'warn' | 'critical'): string {
-	switch (severity) {
-		case 'critical': return COLORS.red;
-		case 'warn': return COLORS.yellow;
-		case 'ok': return COLORS.green;
-	}
-}
-
-function severityIcon(severity: 'ok' | 'warn' | 'critical'): string {
-	switch (severity) {
-		case 'critical': return '!!!';
-		case 'warn': return '!!';
-		case 'ok': return 'OK';
-	}
-}
-
-function formatBar(value: number, max: number, width: number = 20): string {
-	const filled = Math.min(Math.round((value / max) * width), width);
-	const empty = width - filled;
-	const color = value > max * 0.8 ? COLORS.red : value > max * 0.5 ? COLORS.yellow : COLORS.green;
-	return `${color}${'█'.repeat(filled)}${COLORS.dim}${'░'.repeat(empty)}${COLORS.reset}`;
-}
-
-function renderTerminalReport(report: RefactorReport, verbose: boolean): void {
-	console.log();
-	console.log(`${COLORS.cyan}${COLORS.bold}STRUCTURE ANALYSIS${COLORS.reset}`);
-	console.log(`${COLORS.dim}${'─'.repeat(60)}${COLORS.reset}`);
-	console.log(`${COLORS.dim}Target:${COLORS.reset}   ${report.targetPath}`);
-	console.log(`${COLORS.dim}Files:${COLORS.reset}    ${report.totalFiles} discovered, ${report.analyzedFiles} analyzed, ${report.flaggedFiles} flagged`);
-	console.log(`${COLORS.dim}AI:${COLORS.reset}       ${report.aiAnalyzed} files analyzed (${report.cacheHits} from cache)`);
-	console.log(`${COLORS.dim}${'─'.repeat(60)}${COLORS.reset}`);
-	console.log();
-
-	// Show flagged files (or all files if verbose)
-	const toShow = verbose
-		? report.results
-		: report.results.filter(r => r.tier1.severity !== 'ok');
-
-	if (toShow.length === 0) {
-		console.log(`  ${COLORS.green}All files within thresholds. Nothing to flag.${COLORS.reset}`);
-		console.log();
-		return;
-	}
-
-	// Sort by severity (critical first) then by line count
-	toShow.sort((a, b) => {
-		const sevOrder = { critical: 0, warn: 1, ok: 2 };
-		const sevDiff = sevOrder[a.tier1.severity] - sevOrder[b.tier1.severity];
-		if (sevDiff !== 0) return sevDiff;
-		return b.tier1.lineCount - a.tier1.lineCount;
-	});
-
-	for (const result of toShow) {
-		const { tier1, tier2 } = result;
-		const color = severityColor(tier1.severity);
-		const icon = severityIcon(tier1.severity);
-
-		console.log(`  ${color}[${icon}]${COLORS.reset} ${COLORS.bold}${tier1.relativePath}${COLORS.reset}`);
-		console.log(`       ${COLORS.dim}${tier1.language} | ${tier1.lineCount} lines${COLORS.reset} ${formatBar(tier1.lineCount, tier1.hardThreshold)}`);
-		console.log(`       ${COLORS.dim}fn:${tier1.functionCount} exp:${tier1.exportCount} cls:${tier1.classCount} imp:${tier1.importCount}${COLORS.reset}`);
-
-		for (const signal of tier1.signals) {
-			console.log(`       ${color}→ ${signal}${COLORS.reset}`);
-		}
-
-		if (tier2) {
-			const cacheTag = result.cached ? ` ${COLORS.cyan}(cached)${COLORS.reset}` : '';
-			console.log(`       ${COLORS.magenta}AI: ${tier2.responsibilities.length} responsibilities detected${cacheTag}${COLORS.reset}`);
-			for (const r of tier2.responsibilities) {
-				console.log(`       ${COLORS.dim}  • ${r.name}: ${r.description}${COLORS.reset}`);
-			}
-			if (tier2.suggestions.length > 0) {
-				console.log(`       ${COLORS.cyan}Suggested split:${COLORS.reset}`);
-				for (const s of tier2.suggestions) {
-					console.log(`       ${COLORS.cyan}  → ${s.filename}${COLORS.reset} ${COLORS.dim}(${s.responsibilities.join(', ')})${COLORS.reset}`);
-				}
-			}
-			console.log(`       ${COLORS.dim}Effort: ${tier2.effort} | ${tier2.summary.slice(0, 80)}...${COLORS.reset}`);
-		}
-
-		console.log();
-	}
-
-	// Summary
-	console.log(`${COLORS.dim}${'─'.repeat(60)}${COLORS.reset}`);
-	console.log(`${COLORS.bold}SUMMARY${COLORS.reset}`);
-	console.log(`  ${COLORS.red}Critical:${COLORS.reset} ${report.summary.critical}  ${COLORS.yellow}Warnings:${COLORS.reset} ${report.summary.warnings}  ${COLORS.green}OK:${COLORS.reset} ${report.summary.ok}`);
-
-	if (report.summary.topOffenders.length > 0) {
-		console.log(`\n  ${COLORS.bold}Top offenders:${COLORS.reset}`);
-		for (const off of report.summary.topOffenders.slice(0, 5)) {
-			console.log(`    ${COLORS.red}${off.lineCount}${COLORS.reset} lines, ${off.signals} signals — ${off.file}`);
-		}
-	}
-
-	console.log(`\n  ${COLORS.dim}Estimated effort: ${report.summary.estimatedEffort}${COLORS.reset}`);
-	console.log();
-}
-
-function renderJsonReport(report: RefactorReport): void {
-	console.log(JSON.stringify(report, null, 2));
-}
 
 // ─── Main Orchestrator ──────────────────────────────────────────────────────
